Add spatial density model; pull spatial work into DataGenerator class.
RaDecGenerator was doing all of the spatial computation work before,
which made it a weird outlier from the other columns. Lifting that work
into DataGenerator makes the individual columns more reasonable.

This also switches to computing all of the columns for a given box (i.e.
a subset of a chunk) together, which ensures that the contents of data
columns are also identical between "normal" and "edge-only" modes.